requirements.txt
mysql.connecter
clint